October 10 Upd8 (Cleanup Final)
Sorry for the mess of notifs anyone got, I forgot the GitHub CLI was real for a minute and was deleting things by hand because I'm tired & dumb. Anyway, I reorganized the website's files a *lot* & made a few changes to some things. This may not be the only update for today since I want to add some comments to some files so that it's easier for others to read & learn from -& then eventually comment out every file- but I have been working on this during school all day & feel as though it is ready for a big update. Again, sorry if anyone following me got a shit ton of notifs, I am new to doing things like this. ~n1l